My_Weather_Forecast/requirements.txt
fastapi
uvicorn
requests
python-dotenv
apscheduler
pydantic
httpx
loguru
sentry-sdk
pytest 